Change top score selection
Add additional arrays to store the score, and key from the best
choice for each line in the input file, using the line number
as the key for both.

Score array is sorted, then top scoring lines output, decrypted,
along with the line number, key (ord() and chr() value), and
score.